Fix immersed reductions
