+++ src/ImmersedBoundaries/ImmersedBoundaries.jl
@@ -176,7 +176,7 @@ include("grid_fitted_immersed_boundaries.jl")
 include("conditional_fluxes.jl")
 include("conditional_derivatives.jl")
 include("mask_immersed_field.jl")
-include("immersed_fields_reductions.jl")
+include("immersed_reductions.jl")
 
 #####
 ##### Diffusivities (for VerticallyImplicitTimeDiscretization)
